Basic GUI
I made a sample GUI that OBVIOUSLY isn't done, but it has four buttons that all display output when pressed. So here is a SORT of rough implementation of it. The main thing is I think this should already have the name of the zoo at the top of the window. And when entering different "modules" (can change the name) it should already have the specific names of each item (Names, animals, prices, etc.) pertaining to that specific zoo